# src/main/resources/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "https://backend.baruaofficiel.com",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/private/service/documentaires/{id}": {
      "put": {
        "tags": [
          "ServiceDocumentaire"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceDocumentaireRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/private/documents/{id}": {
      "put": {
        "tags": [
          "Document"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServiceDocumentaireRequest": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "integer",
            "format": "int64"
          },
          "organismId": {
            "type": "integer",
            "format": "int64"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "devise": {
            "type": "integer",
            "format": "int64"
          },
          "delayDayOpen": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "delayDayOpen",
          "devise",
          "documentId",
          "organismId",
          "price"
        ]
      },
      "DocumentRequest": {
        "type": "object",
        "properties": {
          "documentTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "documentTypeId",
          "title"
        ]
      },
...
    }
  }
}